# Even faster concretization