Like #5716, this is still not a rework of the concretizer, but it removes a ton of redundant work from our existing concretization algorithm.

Look at the individual commit messages for details.  But, here are the improvements:

Before:
* `xsdk` (30 nodes): 6s
* `dealii` (44 nodes): 10s
* `r-rminer` (134 nodes): 1m38s

After:
* `xsdk`: 1.7s (3.5x)
* `dealii`: 2.5s (4x)
* `r-rminer`: 5s (19x)

- [x] Cache creation of `Compiler` objects from config data
  - this one also prevents Spack from running `xcrun` zillions of times on macs
- [x] Don't redundantly check for providers of non-virtuals when merging dependencies
- [x] Improve `Spec.traverse` so it doesn't construct superfluous `dicts`
- [x] Don't use an `OrderdDict` where we don't have to -- use a `list` (this is the big win for `r-rminer`
- [x] Remove vestigial assertion in `Spec.root` that would ensure that *all* paths led to the root (oops)
- [x] Add a fast path to `satisfies_dependencies` to avoid a bunch of extra work if there are no constraints to satisfy.

`r-rminer` is kind of the upper bound package so far.  Compared to before #5716, concretizing this package is probably 100x faster.  Sorry for all the wasted time everyone! 😳 

@alalazo @adamjstewart @davydden @scheibelp 
@baberlevi (I know you guys do tons of R stuff)