@adamjstewart 2017-10-14
> this one also prevents Spack from running `xcrun` zillions of times on macs

🎉 

> `r-rminer` is kind of the upper bound package so far.

If you want something even harder for testing purposes, there is actually a package that depends on `r-rminer`: https://cran.r-project.org/web/packages/CONDOP/index.html